refactor: rename the SourceSchemaConfig and SourceSchemaType to follow the naming conventions
