+++ packages/ui/src/components/SourceCode/SourceCode.tsx
@@ -5,9 +5,9 @@ import { EditorDidMount } from 'react-monaco-editor';
 import './SourceCode.scss';
 import { SyncButton } from './SyncButton';
 import './workers/enable-workers';
-import { sourceSchemaConfig } from '../../models/camel-entities/SourceSchemaConfig';
+import { sourceSchemaConfig } from '../../models/camel-entities/source-schema-config';
 import { EntitiesContext } from '../../providers/entities.provider';
-import { SourceSchemaType } from '../../models/camel-entities/SourceSchemaType';
+import { SourceSchemaType } from '../../models/camel-entities';
 
 interface SourceCodeProps {
   code: string;
