+++ packages/ui/src/models/camel-entities/source-schema-type.ts
